Draft walker collect children

+++ si-schema-translator/src/main/java/com/octo/keip/schema/model/eip/Occurrence.java
@@ -1,4 +1,4 @@
 package com.octo.keip.schema.model.eip;
 
 // TODO: Handle case when max is unbounded. (e.g. use "unbounded" or -1)
-public record Occurrence(int min, int max) {}
+public record Occurrence(long min, long max) {}
